feat(testMode): add option for utilising testing mode
Closes #133

Have also cleaned up package.json, README.md and test organization.
