+++ src/utils/testModeAPI.js
@@ -1,0 +1,8 @@
+export const gaCalls = [];
+
+export default {
+  calls: gaCalls,
+  ga: (...args) => {
+    gaCalls.push([...args]);
+  }
+};
